feat(Report): Implement report feature

See merge request d5/sm2/training/java/spring-modulith!4

+++ src/main/java/com/rk/WMS/auth/dto/response/LoginResponse.java
@@ -12,7 +12,7 @@ public class LoginResponse {
     private Long userId;
     private String username;
     private String fullName;
-    private Integer warehouseId;
+    private Long warehouseId;
 
     private String accessToken;
     private boolean authenticated;
